backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
torch>=2.2.0
llama-index==0.9.13
google-generativeai==0.3.2
pydantic==2.5.0
numpy>=1.26.0
pandas>=2.0.3
python-multipart==0.0.6
aiofiles==23.2.1
faker==20.1.0
python-dotenv==1.0.0
chromadb>=0.4.0
